requirements.txt
# ===================================
# Core Infrastructure
# ===================================
celery>=5.3.0
redis>=5.0.0
minio>=7.2.0
python-dotenv>=1.0.0

# ===================================
# TTS Module
# ===================================
piper-tts>=1.0.0
pydub>=0.25.1
ffmpeg-python>=0.2.0

# ===================================
# Grammar Module
# ===================================
language-tool-python>=2.8.0

# ===================================
# Translator Module (Future)
# ===================================
# transformers>=4.35.0
# sentencepiece>=0.1.99
# sacremoses>=0.0.53

# ===================================
# Shared Text Processing
# ===================================
nltk>=3.8.0
langdetect>=1.0.9

# ===================================
# CLI & Utilities
# ===================================
click>=8.1.0

# ===================================
# Monitoring
# ===================================
flower>=2.0.0

# ===================================
# Testing
# ===================================
pytest>=7.4.0
pytest-celery>=0.0.0
pytest-cov>=4.1.0
